+++ src/core/src/storage/mod.rs
@@ -137,15 +137,10 @@ pub struct MemStorage {
 }
 
 #[cfg(feature = "branchwater")]
-// Column family for using rocksdb as a Storage
-pub(crate) const STORAGE: &str = "storage";
+pub mod rocksdb;
 
 #[cfg(feature = "branchwater")]
-/// Store data in RocksDB
-#[derive(Debug, Clone)]
-pub struct RocksDBStorage {
-    db: Arc<crate::index::revindex::DB>,
-}
+pub use rocksdb::RocksDBStorage;
 
 pub type Metadata<'a> = BTreeMap<&'a OsStr, &'a piz::read::FileMetadata<'a>>;
 
@@ -662,48 +657,3 @@ impl Storage for MemStorage {
     }
 }
 
-#[cfg(feature = "branchwater")]
-impl RocksDBStorage {
-    pub fn from_path(path: &str) -> Self {
-        let mut opts = crate::index::revindex::RevIndex::db_options();
-        opts.create_if_missing(true);
-        opts.create_missing_column_families(true);
-        opts.prepare_for_bulk_load();
-
-        // prepare column family descriptors
-        let cfs = crate::index::revindex::disk_revindex::cf_descriptors();
-
-        let db =
-            Arc::new(crate::index::revindex::DB::open_cf_descriptors(&opts, path, cfs).unwrap());
-
-        Self { db }
-    }
-
-    pub fn from_db(db: Arc<crate::index::revindex::DB>) -> Self {
-        Self { db: db.clone() }
-    }
-}
-
-#[cfg(feature = "branchwater")]
-impl Storage for RocksDBStorage {
-    fn save(&self, path: &str, content: &[u8]) -> Result<String> {
-        let cf_storage = self.db.cf_handle(STORAGE).unwrap();
-        // TODO(lirber): deal with conflict for path?
-        self.db.put_cf(&cf_storage, path.as_bytes(), &content[..])?;
-        Ok(path.into())
-    }
-
-    fn load(&self, path: &str) -> Result<Vec<u8>> {
-        let cf_storage = self.db.cf_handle(STORAGE).unwrap();
-        let data = self.db.get_cf(&cf_storage, path.as_bytes())?;
-        data.ok_or_else(|| StorageError::DataReadError(path.into()).into())
-    }
-
-    fn args(&self) -> StorageArgs {
-        unimplemented!()
-    }
-
-    fn spec(&self) -> String {
-        "rocksdb://".into()
-    }
-}
